requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
atomicwrites==1.4.0
attrs==21.4.0
backports.zoneinfo==0.2.1
bleach==5.0.0
boto3==1.26.50
botocore==1.29.50
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==36.0.2
cssselect2==0.4.1
dateutils==0.6.12
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.12
django-allauth==0.50.0
django-crispy-forms==1.13.0
django-environ==0.8.1
django-filter==21.1
django-localflavor==3.1
django-model-utils==4.2.0
django-on-heroku==1.1.2
django-rest-auth==0.9.5
django-storages==1.13.2
django-summernote==0.8.20.0
djangorestframework==3.13.1
environ==1.0
fonttools==4.27.1
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
importlib-metadata==4.8.3
iniconfig==1.1.1
jmespath==1.0.1
lxml==4.8.0
Markdown==3.3.6
mercadopago==2.1.0
mime==0.1.0
mysqlclient==2.1.1
oauthlib==3.2.0
oscrypto==1.3.0
packaging==21.3
Pillow==9.4.0
pluggy==1.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pydyf==0.1.2
pyHanko==0.16.0
pyhanko-certvalidator==0.19.8
PyJWT==2.3.0
pyparsing==3.0.8
PyPDF3==1.0.6
pyphen==0.11.0
pytest==7.1.2
pytest-mock==3.7.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-pptx==0.6.21
python-stdnum==1.17
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
svglib==1.5.1
tinycss2==1.1.1
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.1.1
tzdata==2022.7
tzlocal==4.2
uritools==4.0.1
urllib3==1.26.9
validate-docbr==1.10.0
weasyprint==54.3
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.8
XlsxWriter==3.0.3
zipp==3.6.0
zopfli==0.1.9
